requirements.txt
asgiref==3.11.1
asttokens==3.0.1
attrs==26.1.0
colorama==0.4.6
decorator==5.2.1
Django==6.0.4
django-extensions==4.1
djangorestframework==3.17.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
executing==2.2.1
inflection==0.5.1
ipython==9.12.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
matplotlib-inline==0.2.1
parso==0.8.6
prompt_toolkit==3.0.52
pure_eval==0.2.3
Pygments==2.20.0
PyJWT==2.12.1
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.30.0
sqlparse==0.5.5
stack-data==0.6.3
traitlets==5.14.3
typing_extensions==4.15.0
tzdata==2026.1
uritemplate==4.2.0
wcwidth==0.6.0
